# specs/002-multi-user-todo-app/contracts/openapi.yml
openapi: 3.0.0
info:
  title: "Multi-User Todo API"
  version: "1.0.0"
  description: "API for a multi-user todo application."
paths:
  /api/v1/users/signup:
    post:
      summary: "Create a new user"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '200':
          description: "User created successfully"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /api/v1/users/login:
    post:
      summary: "Log in a user"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
      responses:
        '200':
          description: "Login successful"
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  token_type:
                    type: string
  /api/v1/tasks:
    get:
      summary: "Get all tasks for the current user"
      security:
        - bearerAuth: []
      responses:
        '200':
          description: "A list of tasks"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
    post:
      summary: "Create a new task"
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '200':
          description: "Task created successfully"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
  /api/v1/tasks/{task_id}:
    get:
      summary: "Get a specific task"
      security:
        - bearerAuth: []
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: "A single task"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
    put:
      summary: "Update a task"
      security:
        - bearerAuth: []
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: "Task updated successfully"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
    delete:
      summary: "Delete a task"
      security:
        - bearerAuth: []
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: "Task deleted successfully"
  /api/v1/tasks/{task_id}/complete:
    patch:
      summary: "Mark a task as complete or incomplete"
      security:
        - bearerAuth: []
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: "Task completion status updated"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
          format: email
    UserCreate:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    UserLogin:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    Task:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        completed:
          type: boolean
        owner_id:
          type: integer
    TaskCreate:
      type: object
      required:
        - title
      properties:
        title:
          type: string
    TaskUpdate:
      type: object
      required:
        - title
      properties:
        title:
          type: string
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT